requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiohttp-retry==2.8.3
aiosignal==1.3.2
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
billiard==4.2.0
blis==0.7.11
boto3==1.35.81
botocore==1.35.81
Brotli==1.1.0
catalogue==2.0.10
celery==5.4.0
Cerberus==1.3.5
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.41.0
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
constantly==23.10.4
crispy-bootstrap4==2024.1
crispy-tailwind==1.0.3
cron-descriptor==1.4.3
cryptography==43.0.0
cssselect2==0.7.0
cymem==2.0.8
daphne==4.1.2
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==5.0.4
django-address==0.2.8
django-admin-material-dashboard==1.0.10
django-admin-notification==0.2
django-admin-thumbnails==0.2.9
django-allauth==0.61.1
django-autoslug==1.9.9
django-braces==1.15.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cities==0.6.2
django-cities-light==3.10.1
django-ckeditor==6.7.1
django-ckeditor-5==0.2.12
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-countries==7.6.1
django-crispy-forms==2.1
django-crontab==0.7.1
django-embed-video==1.4.4
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
django-form-surveys==2.2.2
django-formtools==2.5.1
django-htmx==1.18.0
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-likert-field==0.2.0
django-localflavor==4.0
django-mail-admin==0.3.2
django-model-utils==4.5.1
django-otp==1.5.4
django-phone-field==1.8.1
django-phonenumber-field==7.3.0
django-private-chat2==1.0.2
django-product-review==0.0.2
django-resized==1.0.2
django-smart-selects==1.7.2
django-social-share==2.3.0
django-star-ratings==0.9.2
django-storages==1.14.4
django-taggit==5.0.1
django-theme==0.3
django-timezone-field==7.0
django-two-factor-auth==1.17.0
django-widget-tweaks==1.5.0
django_paystack==1.0.0
djangochannelsrestframework==1.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
flower==1.1.0
fonttools==4.53.1
frozenlist==1.5.0
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanize==4.10.0
hyperlink==21.0.0
idna==3.7
importlib-metadata==4.13.0
incremental==24.7.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
Js2Py==0.74
jsonfield==3.1.0
kolo==2.33.7
kombu==5.3.7
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mocket==3.12.9
mod_wsgi==5.0.0
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
paypalrestsdk==1.13.3
phonenumbers==8.13.40
pika==1.3.2
pillow==10.3.0
pipwin==0.5.2
platformdirs==4.3.3
preshed==3.0.9
progressbar2==4.4.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.2.1
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg-pool==3.2.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydyf==0.11.0
Pygments==2.18.0
pyjsparser==2.7.1
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyotp==2.9.0
pypaystack2==2.0.3
pyphen==0.16.0
pypng==0.20220715.0
PyPrind==2.11.3
pySmartDL==1.3.4
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-magic==0.4.27
python-magic-bin==0.4.14
python-stdnum==1.20
python-utils==3.8.2
python3-openid==3.2.0
pytz==2024.1
pywin32==306
PyYAML==6.0.2
qrcode==7.4.2
redis==5.0.7
regex==2024.5.15
reportlab==4.2.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.8.0
s3transfer==0.10.4
scikit-learn==1.5.2
scikit-surprise==1.1.4
scipy==1.14.0
service-identity==24.1.0
setuptools==69.5.1
shellingham==1.5.4
simplejwt==2.0.1
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacytextblob==4.0.0
sqlglot==25.21.2
sqlparse==0.5.0
srsly==2.4.8
stripe==4.0.2
swapper==1.3.0
sweetify==2.3.1
tabulate==0.9.0
textblob==0.15.3
thinc==8.2.5
threadpoolctl==3.5.0
tinycss2==1.3.0
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
tweepy==4.14.0
twilio==9.4.1
Twisted==24.7.0
txaio==23.1.1
typer==0.12.5
typing==3.7.4.3
typing_extensions==4.12.0
tzdata==2024.1
tzlocal==5.2
ulid-py==1.1.0
Unidecode==1.3.8
urllib3==2.2.1
uvicorn==0.30.0
vine==5.1.0
waitress==3.0.0
wasabi==1.1.3
watchfiles==0.24.0
wcwidth==0.2.13
weasel==0.4.1
weasyprint==56.1
webencodings==0.5.1
websockets==13.0.1
whitenoise==6.6.0
wrapt==1.16.0
yarl==1.18.3
zipp==3.19.2
zope.interface==7.0.3
zopfli==0.2.3
